front: Change turtle storage to be an object instead of an array
This is needed for receiving the data from the server, as that is given in a hashmap. We also make use a lot of the id for editing, so this helps here as well

Minor changes in here while I was at it: some places expected id to be a number, it is a string now, so I changed that as well. Also I removed the onChange handler from Turtle, it wasn't used anymore anyway. We also now use immutability-helper.